added support for i18n, subpages and static field values
